[feature](selectdb-cloud) Enable multiple BEs test on the same node (#1528)
After supporting smooth upgrading, any added BE on the existing node
will trigger smooth upgrading logic. Developers may not want that
happen when testing multiple BEs on single node. So we add a flag
`IsSmoothUpgrade` to explicitly declare whether the new BE is for
upgrading.

This commit also fix a minor bug that replicas can be hashed to old
BEs. We will not allow that happen by prohibiting any BE with
`IsSmoothUpgrade` flag from being chosen as candidate.

* Update SystemInfoService.java