[feature](selectdb-cloud) Fix coredump when resize the cache size (#1531)
When resize file cache size and restart, file cache will eliminate some segments for fitting the size. And it will encounter an invalid address access error. The reason is that I cache some reference and the datas are removed when eliminating. Fix that by caching key and offset and checking it is valid before used.